moved key functions to own package
