+++ service/resource/worker/service.go
@@ -1,13 +1,12 @@
 package worker
 
 import (
+	"github.com/giantswarm/kvm-operator/service/key"
 	"github.com/giantswarm/kvmtpr"
 	"github.com/giantswarm/microerror"
 	apismetav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	apiv1 "k8s.io/client-go/pkg/api/v1"
-
-	"github.com/giantswarm/kvm-operator/service/resource"
 )
 
 func (s *Service) newService(obj interface{}) (*apiv1.Service, error) {
@@ -26,8 +25,8 @@ func (s *Service) newService(obj interface{}) (*apiv1.Service, error) {
 		ObjectMeta: apismetav1.ObjectMeta{
 			Name: "worker",
 			Labels: map[string]string{
-				"cluster":  resource.ClusterID(*customObject),
-				"customer": resource.ClusterCustomer(*customObject),
+				"cluster":  key.ClusterID(*customObject),
+				"customer": key.ClusterCustomer(*customObject),
 				"app":      "worker",
 			},
 		},
